+++ R/summariseCodeUse.R
@@ -50,7 +50,6 @@ summariseCodeUse <- function(x,
   codeUse <- dplyr::bind_rows(codeUse)
 
 
-
  return(codeUse)
 
 }
@@ -366,6 +365,11 @@ if(length(tableName)>0){
         dplyr::filter(.data$cohort_start_date == !!dplyr::sym(dateName[[1]]))
     }
   }
+
+  if(is.null(codeRecords)){
+    return(NULL)
+  }
+
   codeRecords <- codeRecords %>%
     dplyr::mutate(date = !!dplyr::sym(dateName[[1]])) %>%
     dplyr::mutate(year = lubridate::year(date)) %>%