+++ tests/testthat/test-getICD10StandardCodes.R
@@ -1,5 +1,5 @@
 test_that("db without icd10 codes loaded", {
-  backends <- c("database", "arrow", "data_frame")
+  backends <- c("database", "data_frame")
   for (i in seq_along(backends)) {
     cdm <- mockVocabRef(backend = backends[i])
     codes <- getICD10StandardCodes(cdm = cdm,